feat: refactor networking util
