+++ test/e2e/network/BUILD
@@ -62,6 +62,7 @@ go_library(
         "//test/e2e/framework/endpoints:go_default_library",
         "//test/e2e/framework/ingress:go_default_library",
         "//test/e2e/framework/log:go_default_library",
+        "//test/e2e/framework/networking:go_default_library",
         "//test/e2e/framework/providers/gce:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
         "//test/e2e/network/scale:go_default_library",
